split files, refactor
