+++ src/jacdac.ts
@@ -1,3 +1,4 @@
-require('./pxtutils')
-require('./jd')
+require('./jdutils')
+require('./jdpacket')
+require('./jddevice')
 require('./jdpretty')
